add counting of attempts by activity-id to the ActivityCorrelator
